requirements.txt
# Core Environment
python-dotenv>=1.0.1

# LangChain Ecosystem (Updated for 3.13 compatibility)
langchain-core>=0.3.30
langchain-google-genai>=2.0.9
langchain>=0.3.14

# Data Validation (Crucial for 3.13)
pydantic>=2.10.0

# Integration & Utilities
todoist-api-python>=2.1.7
tzdata>=2024.2